foram adicionados argumentos para desabilitar funcionalidades
related #27

+++ helpers/restfy.py
@@ -1,10 +1,11 @@
 import json
-from django.http.response import HttpResponse
+from django.http.response import HttpResponse, HttpResponseNotAllowed
 from django.db import transaction
 from django.db.models import Q
 
 
-def make_rest(Serializer):
+def make_rest(Serializer, allow_create=True, allow_list=True,
+         allow_get=True, allow_update=True, allow_delete=True):
 
     Model = Serializer.Model()
 
@@ -187,24 +188,40 @@ def make_rest(Serializer):
 
     def _index(request):
         response = None
-        if request.method == 'GET':
+        if allow_list and request.method == 'GET':
             response = _list(request)
 
-        elif request.method == 'POST':
+        elif allow_create and request.method == 'POST':
             response = _create(request)
+        else:
+            permitted_methods = []
 
+            allow_list and permitted_methods.append('GET')
+            allow_create and permitted_methods.append('POST')
+            
+            response = HttpResponseNotAllowed(
+                permitted_methods = permitted_methods
+            )
 
         return response
 
     def _by_id(request, id):
-            if request.method == 'GET':
+            if allow_get and request.method == 'GET':
                 return _get_by_id(request, id)
-            elif request.method == 'DELETE':
+            elif allow_delete and request.method == 'DELETE':
                 return _delete_by_id(request, id)
-            elif request.method == 'PUT':
+            elif allow_update and request.method == 'PUT':
                 return _update_by_id(request, id)
             else:
-                return HttpResponse(status=501)
+                permitted_methods = []
+
+                allow_get and permitted_methods.append('GET')
+                allow_delete and permitted_methods.append('DELETE')
+                allow_update and permitted_methods.append('UPDATE')
+            
+                return HttpResponseNotAllowed(
+                permitted_methods = permitted_methods
+                )
 
     
     return _index, _by_id
